lib: Decoupled require_all from main index
